+++ flow-typed/micro-memoize.js
@@ -6,6 +6,7 @@ declare module 'micro-memoize' {
 
   declare type Options = {
     isEqual?: (firstValue: any, secondValue: any) => boolean,
+    isMatchingKey?: (cacheKey: Array<any>, key: Array<any>) => boolean,
     isPromise?: boolean,
     maxSize?: number,
     onCacheAdd?: (cache: Cache, options: Options) => void,