+++ index.d.ts
@@ -5,6 +5,7 @@ interface Cache = {
 
 interface Options {
   isEqual?: (firstValue: any, secondValue: any) => boolean;
+  isMatchingKey?: (cacheKey: Array<any>, key: Array<any>) => boolean;
   isPromise?: boolean;
   maxSize?: number;
   onCacheAdd?: (cache: Cache, options: Options, memoized: Function) => void;